+++ R/mod_country_comparison.R
@@ -45,7 +45,6 @@ mod_country_comparison_server <- function(input, output, session, orig_data){
           aggregate_province_timeseries_data() %>%
           filter(Country.Region %in% input$select_countries) %>%
           filter(contagion_day > 0) %>%
-          select(-ends_with("rate")) %>%
           arrange(desc(date))
       })