+++ R/utils.R
@@ -93,6 +93,14 @@ new_case_colors <- c(
 #' Color Palette
 #'
 #' @export
+rate_colors <- c(
+  "growth_factor" = "#dd4b39",
+  "death_rate" = "black"
+)
+
+#' Color Palette
+#'
+#' @export
 new_total_colors <- c(
   "total" = "#C8C8C8",
   "new" = "#ea8b5b"
@@ -120,6 +128,13 @@ load_countries_data <- function(destpath = system.file("./countries_data", packa
                               verbose = FALSE)
 }
 
+#' Sort type by max
+#'
+#' @param data data to be sorted
+sort_type_by_max <- function(data) {
+  c("active", "recovered", "deaths") %>% .[order(sapply(data[.], max))]
+}
+
 #' Round up to the next decine
 #' Ref: https://stackoverflow.com/questions/6461209/how-to-round-up-to-the-nearest-10-or-100-or-x
 #' @param x number to round
@@ -127,3 +142,50 @@ load_countries_data <- function(destpath = system.file("./countries_data", packa
 #' @export
 roundUp <- function(x) 10^ceiling(log10(x))
 
+#' align countries names between population dataframe and orig_data dataframe
+#' @param data data.frame
+#' @export
+align_country_names_pop <- function(data){
+  data$Country.Region <- data$Country.Region %>%
+    recode(
+      #Note Congo (Brazzaville), Diamond Proncess, Guadeloupe, Martinique, Reunion,  present in orig_data not in population; and Tristan da Cunha(UK)  present in population, not in orig_data
+
+      "Antigua and Barbuda" = "Antigua and Barb.",
+      "Bosnia and Herzegovina" = "Bosnia and Herz.",
+      "Cape Verde" = "Cabo Verde",
+      "Cote d'Ivoire" = "C\\u00f4te d'Ivoire",
+      "Czech Republic" = "Czechia",
+      "Dominican Republic" = "Dominican Rep.",
+      "Eswatini" = "eSwatini",
+      "French Guiana" = "Guyana",
+      "North Macedonia" = "Macedonia",
+      "UK" = "United Kingdom",
+      "USA" = "United States of America",
+      "Vatican City" = "Vatican"
+    )
+  data
+}
+
+#' align countries names between population dataframe and orig_data dataframe
+#' @param data data.frame
+#' @export
+align_country_names_pop_reverse <- function(data){
+  data$Country.Region <- data$Country.Region %>%
+    recode(
+      #Note Congo (Brazzaville), Diamond Proncess, Guadeloupe, Martinique, Reunion,  present in orig_data not in population; and Tristan da Cunha(UK)  present in population, not in orig_data
+
+      "Antigua and Barb." = "Antigua and Barbuda",
+      "Bosnia and Herz." = "Bosnia and Herzegovina",
+      "Cabo Verde" = "Cape Verde",
+      "C\\u00f4te d'Ivoire" = "Cote d'Ivoire",
+      "Czechia" = "Czech Republic",
+      "Dominican Rep." = "Dominican Republic",
+      "eSwatini" = "Eswatini",
+      "Guyana" = "French Guiana",
+      "Macedonia" = "North Macedonia",
+      "United Kingdom" = "UK",
+      "United States of America" = "USA",
+      "Vatican" = "Vatican City"
+    )
+  data
+}